vllm/requirements.txt
vllm>=0.6.0
transformers>=4.40.0
torch>=2.1.0
accelerate
python-dotenv
